requirements.txt
bokeh==2.1.1
transformers>=2.11.0
sentence_transformers>=0.2.6.2
scikit-learn==0.22.1
pandas==1.0.5
streamlit==0.62.0
scipy==1.4.1
nltk==3.5
emoji==0.5.4
pytest==5.4.3
numpy==1.18.5
seaborn==0.10.1
GitPython==3.1.3
protobuf==3.12.2
spacy==2.3.0
tweepy==3.8.0
vaderSentiment==3.3.2
flake8>=3.7.9
google-cloud-storage==1.29.0
pytest-custom-exit-code==0.3.0
twarc>=1.10.0
pyarrow==0.17.1
gcsfs==0.6.2
numba==0.50.1
swifter==0.305
parso<0.8.0,>=0.7.0
tensorflow==2.3.0
Pillow>=7.1.0